back de polla, creo q ya esta listo

+++ backend/motoextreme/motoextreme/src/main/java/com/motoextreme/motoextreme/dtos/response/CarritoItemResponseDTO.java
@@ -2,14 +2,14 @@ package com.motoextreme.motoextreme.dtos.response;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CarritoItemResponseDTO {
-
     private Long id;
     private int cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
-
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
     private Long motoId;
     private Long accesorioId;
 }
